requirements.txt
boto3
pandas
numpy
streamlit-audiorecorder
datetime
pdfkit
streamlit-quill